[examples] update Biology example with new interface
