+++ examples/Biology/StepSystem/StepSystemReduced.cpp
@@ -1,6 +1,6 @@
 #include "SiconosKernel.hpp"
 #include "const.h"
-#include "NonlinearRelationReduced2.h"
+#include "NonlinearRelationReduced2.hpp"
 #include "myDS.h"
 #include <stdio.h>
 #include <stdlib.h>
@@ -74,9 +74,9 @@ int main(int argc, char *argv[])
   SP::Interaction aI(new Interaction(aNSL,aR));
 
 //****BUILD THE SYSTEM
-  SP::Model  aM(new Model(0,sTf));
-  aM->nonSmoothDynamicalSystem()->insertDynamicalSystem(aDS);
-  aM->nonSmoothDynamicalSystem()->link(aI,aDS);
+  SP::NonSmoothDynamicalSystem  aN(new NonSmoothDynamicalSystem(0,sTf));
+  aN->insertDynamicalSystem(aDS);
+  aN->link(aI,aDS);
 
 // -- (1) OneStepIntegrators --
   SP::OneStepIntegrator  aEulerMoreauOSI ;
@@ -94,15 +94,12 @@ int main(int argc, char *argv[])
   osnspb->setNumericsVerboseMode(0);
 
 // -- (4) Simulation setup with (1) (2) (3)
-  SP::TimeStepping aS(new TimeStepping(aTD,aEulerMoreauOSI,osnspb));
+  SP::TimeStepping aS(new TimeStepping(aN, aTD, aEulerMoreauOSI,osnspb));
   aS->setComputeResiduY(true);
   aS->setUseRelativeConvergenceCriteron(false);
-
-// Initialization
-  printf("-> Initialisation \n");
-  aM->setSimulation(aS);
-  aM->initialize();
-  printf("-> End of initialization \n");
+  aS->setNewtonTolerance(1e-4);
+  aS->setNewtonMaxIteration(200);
+  aS->setResetAllLambda(false);
 
 // BUILD THE STEP INTEGRATOR
 
@@ -122,7 +119,7 @@ int main(int argc, char *argv[])
 
   printf("=== Start of simulation: %d steps ===  \n", NBStep);
 
-  dataPlot(0, 0) = aM->t0();
+  dataPlot(0, 0) = aN->t0();
   dataPlot(0,1) = x->getValue(0);
   dataPlot(0,2) = x->getValue(1);
   dataPlot(0, 3) = lambda->getValue(0);
@@ -141,8 +138,7 @@ int main(int argc, char *argv[])
     std::cout<<"-> Running step:"<<k<<std::endl;
 #endif
     cmp++;
-
-    aS->newtonSolve(1e-4, 200);
+    aS->advanceToEvent();
  
     dataPlot(cmp, 0) = aS->nextTime();
     dataPlot(cmp, 1) = x->getValue(0);
