+++ examples/Biology/StepSystem/src/NonlinearRelationReduced.hpp
@@ -10,9 +10,6 @@ public:
   NonlinearRelationReduced();
   virtual ~NonlinearRelationReduced() {};
 
-  virtual void initializeWorkVectorsAndMatrices(Interaction& inter, VectorOfBlockVectors& DSlink, VectorOfVectors& workV, VectorOfSMatrices& workM);
-
-
   /** default function to compute h
    *  \param double : current time
    */
