+++ examples/Biology/StepSystem/src/NonlinearRelationWithSignInversed.hpp
@@ -10,9 +10,6 @@ public:
   NonlinearRelationWithSignInversed();
   virtual ~NonlinearRelationWithSignInversed() {};
 
-  virtual void initializeWorkVectorsAndMatrices(Interaction& inter, VectorOfBlockVectors& DSlink, VectorOfVectors& workV, VectorOfSMatrices& workM);
-
-
   /** default function to compute h
    *  \param double : current time
    */
